+++ app/src/main/java/com/example/easyprice/data/FavoritesManager.kt
@@ -8,7 +8,6 @@ object FavoritesManager {
     val favorites = mutableStateListOf<Product>()
 
     fun add(product: Product) {
-        // Buscamos el producto por sus propiedades, no por el objeto en sí
         val existingProduct = favorites.find { it.name == product.name && it.price == product.price }
         if (existingProduct != null) {
             val index = favorites.indexOf(existingProduct)
@@ -22,7 +21,6 @@ object FavoritesManager {
     }
 
     fun remove(product: Product) {
-        // Buscamos el producto a eliminar por sus propiedades para asegurar que se encuentre
         val productToRemove = favorites.find { it.name == product.name && it.price == product.price }
         if (productToRemove != null) {
             favorites.remove(productToRemove)
@@ -30,7 +28,6 @@ object FavoritesManager {
     }
 
     fun increaseQuantity(product: Product) {
-        // Buscamos el producto por sus propiedades para encontrar el índice
         val index = favorites.indexOfFirst { it.name == product.name && it.price == product.price }
         if (index != -1) {
             val currentProduct = favorites[index]
@@ -40,11 +37,9 @@ object FavoritesManager {
     }
 
     fun decreaseQuantity(product: Product) {
-        // Buscamos el producto por sus propiedades para encontrar el índice
         val index = favorites.indexOfFirst { it.name == product.name && it.price == product.price }
         if (index != -1) {
             val currentProduct = favorites[index]
-            // Solo disminuimos si la cantidad es mayor que 1
             if (currentProduct.quantity > 1) {
                 val updatedProduct = currentProduct.copy(quantity = currentProduct.quantity - 1)
                 favorites[index] = updatedProduct
@@ -52,7 +47,8 @@ object FavoritesManager {
         }
     }
 
-    fun total(): Int {
+    // Se actualiza a Double para permitir decimales en el total
+    fun total(): Double {
         return favorites.sumOf { it.price * it.quantity }
     }
 }
